Add files via upload
The Anisotropic of Induced Current Density (ACID) map for the azulene molecule in its ground state is presented in the accompanying movie, which explores the transannular bond from various rotational angles. The map reveals an isosurface at the transannular bond that exhibits characteristics intermediate between sigma and pi bonding, allowing a small electron flux of diatropic current along this bond.